Removed chart.height and chart.cols functions
Because there are always 12 months in a year, there isn't really a need for chart.cols(). In addition, chart.height() isn't really used. However, I believe chart.width() should stay because some users of the library may wish to make a version of the site with a bigger or smaller version of the squares, and in order to do so, chart.width would need to be supported. For our use case, keeping it at 200 is good, but for extensibility allowing it to be changed is even better.